requirements.txt
azure-ai-vision-imageanalysis==1.0.0
python-dotenv==1.1.0
sqlmodel==0.0.24
alembic==1.15.1
asyncpg==0.30.0
bcrypt==4.3.0
colorlog==6.9.0
pydantic-settings==2.8.1
PyJWT==2.10.1
python-multipart==0.0.20
SQLAlchemy==2.0.40
sqlmodel==0.0.24
passlib~=1.7.4
fastapi[standard]==0.115.12
pydantic==2.11.0
azure-storage-blob==12.25.1
azure-identity==1.21.0
